TesterRunner/requirements.txt
defusedxml==0.5.0
diff-match-patch==20181111
Django==2.1.1
django-crispy-forms==1.7.2
django-filter==2.1.0
django-formtools==2.1
django-import-export==1.2.0
django-reversion==3.0.3
django-tastypie==0.14.2
djangorestframework==3.9.2
et-xmlfile==1.0.1
jdcal==1.4
odfpy==1.4.0
openpyxl==2.6.1
PyMySQL==0.9.3
python-dateutil==2.8.0
python-mimeparse==1.6.0
pytz==2018.9
PyYAML==3.13
six==1.12.0
tablib==0.12.1
unicodecsv==0.14.1
xlrd==1.2.0
xlwt==1.3.0
